+++ src/main/java/baseball/service/RoundResultConsumer.java
@@ -1,6 +1,6 @@
 package baseball.service;
 
-import baseball.domain.BaseballRoundResult;
+import baseball.common.dto.BaseballRoundResult;
 
 @FunctionalInterface
 public interface RoundResultConsumer {
