+++ src/test/java/baseball/domain/BaseballNumbersTest.java
@@ -1,11 +1,12 @@
 package baseball.domain;
 
+import baseball.common.dto.BaseballRoundResult;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 class BaseballNumbersTest {
